codegen: narrow GCC 13 sanitized mask and probe conversions
The GCC 13 Debug --sanitize build rejects 27 -Wconversion and
-Wsign-conversion diagnostics in codegen.c and in the sources machine.c
includes. Make each narrowing explicit where its bound is visible:

- u64 callee-saved and allocatable mask bits accumulated into u32
  counters: cast the completed `& 1u` addend, which is 0 or 1.
- u16 identity-plan and preserved-vector masks, whose shift promotes to
  int before `& 1u`: promote the mask itself to u32, so the shift and the
  mask stay unsigned.
- conditionals whose u8/u16 arms promote to int: narrow the completed
  conditional once. The byte probes keep their low-byte truncation, and
  the variant mask is at most 0xffff because variants_in_range bounds the
  count to 1..16.

Every changed expression yields the same value for every input, and no
warning, -Werror or sanitizer policy changes. Advances #571; #562 stays
open.

+++ src/buster/lib/compiler/codegen/machine_select.c
@@ -122,7 +122,7 @@ bool machine_selection_assembly_identity_plan(IrProgram* program, IrFunction* fu
             if (selected && matching)
             {
                 u32 match = IR_INLINE_ASSEMBLY_CONSTRAINT_MATCH_INDEX(constraint);
-                selected = match < index && ((plan->outputs >> match) & 1u) && !((plan->read_write >> match) & 1u) &&
+                selected = match < index && (((u32)plan->outputs >> match) & 1u) && !(((u32)plan->read_write >> match) & 1u) &&
                            plan->source_operands[match] == UINT8_MAX && plan->sizes[match] == type->layout.size;
                 if (selected)
                 {
